refactor tick mark utils
closes #445
